+++ web/views/search_view.ex
@@ -18,6 +18,7 @@ defmodule Bep.SearchView do
 
   def render("scripts.results.html", _assigns) do
     ~s{<script>require("web/static/js/results_lazy_loading")</script>}
+    <> ~s{<script>require("web/static/js/results_filter")</script>}
     |> raw
   end
